Changes on Node.Builder + Context logic
Fix components id to int.
Remove the possibility for Node.Builder to add nodes as children.
Continue Context logic + comments

+++ Ignite/Node_Components.cs
@@ -18,18 +18,18 @@ namespace Ignite
         /// <summary>
         /// Trigger when a component is replaced, send the old component id and the component as payload
         /// </summary>
-        public event Action<ulong, IComponent>? OnComponentReplaced;
+        public event Action<int, IComponent>? OnComponentReplaced;
 
         /// <summary>
         /// Trigger when a component is removed, send the component id and the whether it was from node deletion or not as payload
         /// </summary>
-        public event Action<ulong, bool>? OnComponentRemoved;
+        public event Action<int, bool>? OnComponentRemoved;
 
         /// <summary>
         /// Collection of components referenced by there id
         /// </summary>
-        public Dictionary<ulong, IComponent> Components { get; protected set; } =
-            new Dictionary<ulong, IComponent>();
+        public Dictionary<int, IComponent> Components { get; protected set; } =
+            new Dictionary<int, IComponent>();
 
         private readonly ComponentLookupTable _lookup;
 
@@ -57,7 +57,7 @@ namespace Ignite
         /// <summary>
         /// Check whether the node has a component from it's index or not
         /// </summary>
-        internal bool HasComponent (ulong index )
+        internal bool HasComponent (int index )
         {
             return Components.ContainsKey(index);
         }
@@ -134,7 +134,7 @@ namespace Ignite
         /// <summary>
         /// Get the index of a component of <see cref="Type"/> <typeparamref name="T"/>
         /// </summary>
-        internal ulong GetComponentIndex<T> ()
+        internal int GetComponentIndex<T> ()
         {
             return _lookup[typeof(T)];
         }
@@ -142,7 +142,7 @@ namespace Ignite
         /// <summary>
         /// Get the index of a component of <see cref="Type"/> <paramref name="type"/>
         /// </summary>
-        internal ulong GetComponentIndex ( Type type )
+        internal int GetComponentIndex ( Type type )
         {
             return _lookup[type];
         }
@@ -204,7 +204,7 @@ namespace Ignite
         /// </summary>
         public Node AddOrReplaceComponent<T>(T component) where T : class, IComponent
         {
-            ulong index = _lookup[component.GetType()];
+            int index = _lookup[component.GetType()];
             if (Components.ContainsKey(index))
             {
                 Components[index] = component;
@@ -228,7 +228,7 @@ namespace Ignite
         /// </summary>
         public Node RemoveComponent<T> ()
         {
-            ulong index = _lookup[typeof(T)];
+            int index = _lookup[typeof(T)];
             Components.Remove(index);
             OnComponentRemoved?.Invoke(index, false);
             return this;
@@ -239,7 +239,7 @@ namespace Ignite
         /// </summary>
         public Node RemoveComponent ( Type type )
         {
-            ulong index = _lookup[type];
+            int index = _lookup[type];
             Components.Remove(index);
             OnComponentRemoved?.Invoke(index, false);
             return this;
